requirements.txt
fastapi
uvicorn[standart]
sqlalchemy
asyncpg
alembic
pydantic
python-dotenv
pydantic[email]
pydantic_settings
python-jose
python-multipart
websockets
aiohttp
aioconsole